This is very broken, but committing to get it off my computer
I don't know if this will actually run in this state, but there are changes that should be preserved. Postman released an update to their API to include an import by JSON that I intend to utilize. I started to make the changes necessary to use the PostmanAPI calls only.